// backOffice/src/app/core/models/form-field-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface FormFieldDto {
  label?: string;
  order?: number;
  placeholder?: string;
  rangeEnd?: number;
  rangeStart?: number;
  regex?: string;
  regexErrorMessage?: string;
  required?: boolean;
  selectOptions?: Array<string>;
  type?: string;
}